refactor: separating code into more modules
