+++ src/Base/Base.hs
@@ -12,9 +12,10 @@ import Data.Time (UTCTime (..), addDays)
 import Database.Esqueleto.Experimental (Key, OrderBy, PersistField (..), SqlExpr, Value (..), asc, count, delete, desc, from, fromSqlKey, get, getBy, groupBy, innerJoin, insert, insertMany, insertMany_, insert_, just, leftJoin, like, limit, offset, on, orderBy, replace, select, table, unionAll_, val, where_, withRecursive, (%), (&&.), (++.), (:&) (..), (<.), (==.), (>=.), (?.), (^.), (||.))
 import Database.Persist.Postgresql (ConnectionString, Entity (..), rawExecute, toSqlKey, withPostgresqlConn)
 import Database.Persist.Sql (SqlPersistT, runMigration, runSqlConn)
-import Handlers.Base (Limit (..), Offset (..), Success (..))
+import Handlers.Base.Base (Limit (..), Offset (..), Success (..))
+import Handlers.Web.Web (CategoryInternal (..), NewsEditInternal (..), NewsInternal (..), NewsOut (..), UserInternal (..))
 import Scheme (Category (..), ColumnType (..), EntityField (..), FilterItem (..), Find (..), Image (..), ImageBank (..), News (..), Password (..), SortOrder (..), Unique (..), User (..), migrateAll)
-import Types (CategoryInternal (..), Content (..), Label (..), Login (..), Name (..), NewsEditInternal (..), NewsInternal (..), NewsOut (..), NumberImage (..), PasswordUser (..), Title (..), URI_Image (..), UserInternal (..))
+import Types (Content (..), Label (..), Login (..), Name (..), NumberImage (..), PasswordUser (..), Title (..), URI_Image (..))
 
 --
 makeAndFillTables :: ConnectionString -> IO ()
@@ -71,10 +72,10 @@ validCopyRight connString login title = do
         fetchUserFromNews :: (MonadIO m) => SqlPersistT m [Value T.Text]
         fetchUserFromNews = select $ do
           (news :& user) <-
-            from $
-              table @News
+            from
+              $ table @News
                 `innerJoin` table @User
-                  `on` (\(n :& u) -> n ^. NewsUserId ==. (u ^. UserId))
+              `on` (\(n :& u) -> n ^. NewsUserId ==. (u ^. UserId))
           where_ (news ^. NewsTitle ==. (val . getTitle) title)
           pure (user ^. UserLogin)
 
@@ -91,10 +92,10 @@ validPassword connString login password = do
     fetchSaltAndPassword :: (MonadFail m, MonadIO m) => SqlPersistT m [Value T.Text]
     fetchSaltAndPassword = select $ do
       (user :& pass) <-
-        from $
-          table @User
+        from
+          $ table @User
             `innerJoin` table @Password
-              `on` (\(u :& p) -> u ^. UserPasswordId ==. p ^. PasswordId)
+          `on` (\(u :& p) -> u ^. UserPasswordId ==. p ^. PasswordId)
       where_ (user ^. UserLogin ==. (val . getLogin) login)
       pure (pass ^. PasswordQuasiPassword)
 
@@ -110,14 +111,14 @@ pullAllNews connString configLimit userOffset userLimit columnType sortOrder mbF
           unValue
           ( select $ do
               (news :& author :& categoryName :& imageBank) <-
-                from $
-                  table @News
+                from
+                  $ table @News
                     `innerJoin` table @User
-                      `on` (\(n :& a) -> n ^. NewsUserId ==. a ^. UserId)
+                  `on` (\(n :& a) -> n ^. NewsUserId ==. a ^. UserId)
                     `innerJoin` table @Category
-                      `on` (\(n :& _ :& c) -> n ^. NewsCategoryId ==. c ^. CategoryId)
+                  `on` (\(n :& _ :& c) -> n ^. NewsCategoryId ==. c ^. CategoryId)
                     `leftJoin` table @ImageBank
-                      `on` (\(n :& _ :& _ :& ib) -> just (n ^. NewsId) ==. ib ?. ImageBankNewsId)
+                  `on` (\(n :& _ :& _ :& ib) -> just (n ^. NewsId) ==. ib ?. ImageBankNewsId)
               groupBy (news ^. NewsTitle, news ^. NewsCreated, author ^. UserName, categoryName ^. CategoryLabel, imageBank ?. ImageBankNewsId)
               -- search substring
               maybe
@@ -193,20 +194,20 @@ fetchFullNews configLimit userLimit title = do
     fetchLabel :: (MonadIO m) => SqlPersistT m [Entity Category]
     fetchLabel = select $ do
       (news :& category) <-
-        from $
-          table @News
+        from
+          $ table @News
             `innerJoin` table @Category
-              `on` (\(n :& c) -> n ^. NewsCategoryId ==. (c ^. CategoryId))
+          `on` (\(n :& c) -> n ^. NewsCategoryId ==. (c ^. CategoryId))
       where_ (news ^. NewsTitle ==. (val . getTitle) title)
       pure category
 
     fetchUser :: (MonadIO m) => SqlPersistT m [Entity User]
     fetchUser = select $ do
       (news :& user) <-
-        from $
-          table @News
+        from
+          $ table @News
             `innerJoin` table @User
-              `on` (\(n :& c) -> n ^. NewsUserId ==. (c ^. UserId))
+          `on` (\(n :& c) -> n ^. NewsUserId ==. (c ^. UserId))
       where_ (news ^. NewsTitle ==. (val . getTitle) title)
       pure user
 
@@ -233,12 +234,12 @@ fetchFullNews configLimit userLimit title = do
     fetchActionImage :: (MonadIO m) => SqlPersistT m [Entity Image]
     fetchActionImage = select $ do
       (news :& _imagebank :& image) <-
-        from $
-          table @News
+        from
+          $ table @News
             `innerJoin` table @ImageBank
-              `on` (\(n :& i) -> n ^. NewsId ==. (i ^. ImageBankNewsId))
+          `on` (\(n :& i) -> n ^. NewsId ==. (i ^. ImageBankNewsId))
             `innerJoin` table @Image
-              `on` (\(_ :& i :& im) -> (i ^. ImageBankImageId) ==. (im ^. ImageId))
+          `on` (\(_ :& i :& im) -> (i ^. ImageBankImageId) ==. (im ^. ImageId))
       where_ (news ^. NewsTitle ==. (val . getTitle) title)
       limit (fromIntegral . min configLimit . getLimit $ userLimit)
       pure image
