+++ src/Handlers/Web/Api.hs
@@ -1,0 +1,6 @@
+module Handlers.Web.Api (endPointNews, endPointUsers, endPointImages, endPointCategories) where
+
+import Handlers.Web.Category.CategoryApi (endPointCategories)
+import Handlers.Web.Image.ImageApi (endPointImages)
+import Handlers.Web.News.NewsApi (endPointNews)
+import Handlers.Web.User.UserApi (endPointUsers)
