+++ src/Handlers/Web/Category/CategoryApi.hs
@@ -1,0 +1,45 @@
+-- {-# LANGUAGE DataKinds #-}
+module Handlers.Web.Category.CategoryApi (endPointCategories) where
+
+import qualified Data.Text as T
+import qualified Data.Text.Encoding as E
+import qualified Handlers.Base.Base
+import qualified Handlers.Logger
+import Handlers.Web.Category.Create.Api (createCategory)
+import Handlers.Web.Category.Get.Api (existingCategories)
+import Handlers.Web.Category.Update.Api (updateCategory)
+import Handlers.Web.Web (Client (..), Handle (..))
+import Network.Wai (Request, Response, queryString, rawPathInfo)
+import Web.WebType (queryToPanigate)
+
+endPointCategories :: (Monad m) => Handle m -> Request -> m Response
+endPointCategories h req = do
+  let logHandle = logger h
+      baseHandle = base h
+  Handlers.Logger.logMessage logHandle Handlers.Logger.Debug "end Point Categories"
+  Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (E.decodeUtf8 $ rawPathInfo req)
+  case rawPathInfo req of
+    "/categories/create" -> do
+      case client h of
+        Client (Just adminRole) _ _ -> createCategory adminRole h req -- create a category for only admin
+        _ -> do
+          Handlers.Logger.logMessage logHandle Handlers.Logger.Warning "Access denied"
+          pure (response404 h) -- todo. replace 404 for another error
+    "/categories/edit" -> do
+      case client h of
+        Client (Just adminRole) _ _ -> updateCategory adminRole h req -- edit a category for only admin
+        _ -> do
+          Handlers.Logger.logMessage logHandle Handlers.Logger.Warning "Access denied"
+          pure (response404 h)
+    "/categories" -> do
+      let queryLimit = queryString req
+          (userOffset, userLimit) = queryToPanigate queryLimit
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug "Query String:"
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (T.pack $ show queryLimit)
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (T.pack $ show (userOffset, userLimit))
+
+      let newBaseHandle = baseHandle {Handlers.Base.Base.userOffset = userOffset, Handlers.Base.Base.userLimit = userLimit}
+      existingCategories (h {base = newBaseHandle}) req
+    _ -> do
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Warning "End point not found"
+      pure $ response404 h
