+++ src/Handlers/Web/Category/Types.hs
@@ -1,0 +1,8 @@
+module Handlers.Web.Category.Types (CategoryInternal (..)) where
+
+import Types (Label (..))
+
+data CategoryInternal = CategoryInternal
+  { labelCategory :: Label,
+    parentCategory :: Maybe Label
+  }
