+++ src/Handlers/Web/Image/Types.hs
@@ -1,0 +1,1 @@
+module Handlers.Web.Image.Types where
