+++ src/Handlers/Web/News/NewsApi.hs
@@ -1,0 +1,89 @@
+-- {-# LANGUAGE DataKinds #-}
+module Handlers.Web.News.NewsApi (endPointNews) where
+
+import qualified Data.Text as T
+import qualified Data.Text.Encoding as E
+import qualified Handlers.Base.Base
+import qualified Handlers.Logger
+import Handlers.Web.News.Create.Api (createNews)
+import Handlers.Web.News.Get.Api (existingNews)
+import Handlers.Web.News.Update.Api (updateNews)
+import Handlers.Web.Web (Client (..), Handle (..))
+import Network.HTTP.Types (Query)
+import Network.Wai (Request, Response, queryString, rawPathInfo)
+import Scheme (FilterItem (FilterPublishOrAuthor))
+import Types (Login (..))
+import Web.WebType (queryToFilters, queryToFind, queryToPanigate, queryToSort)
+
+endPointNews :: (Monad m) => Handle m -> Request -> m Response
+endPointNews h req = do
+  let logHandle = logger h
+  Handlers.Logger.logMessage logHandle Handlers.Logger.Debug "end Point News"
+  Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (E.decodeUtf8 $ rawPathInfo req)
+  case rawPathInfo req of
+    "/news/create" -> do
+      case client h of
+        Client _ (Just publisherRole) _ -> createNews publisherRole h req -- create news for only publisher
+        _ -> do
+          Handlers.Logger.logMessage logHandle Handlers.Logger.Warning "Access denied"
+          pure (response404 h) --
+    "/news/edit" ->
+      case client h of
+        Client _ _ (Just author_) -> do
+          updateNews author_ h req -- edit news for only author
+        _ -> do
+          Handlers.Logger.logMessage logHandle Handlers.Logger.Warning "Access denied"
+          pure (response404 h) --
+    "/news" -> do
+      -- get all news
+      let queryLimit = queryString req
+          (userOffset, userLimit) = queryToPanigate queryLimit
+          sortWeb = queryToSort queryLimit
+          findWeb = queryToFind queryLimit
+          filtersWeb = queryToFilters queryLimit
+      -- Debug
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug "Query String:"
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (T.pack $ show queryLimit)
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (T.pack $ show (userOffset, userLimit))
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (T.pack $ show sortWeb)
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Debug (T.pack $ show findWeb)
+      let filterPublishOrAuthor = FilterPublishOrAuthor (fmap getLogin $ author $ client h)
+      mapM_
+        (Handlers.Logger.logMessage logHandle Handlers.Logger.Debug . T.pack . show)
+        (filterPublishOrAuthor : filtersWeb)
+      existingNews (foldSets queryLimit h [setFilters, setFind, setSort, setPanigate]) req
+    _ -> do
+      Handlers.Logger.logMessage logHandle Handlers.Logger.Warning "End point not found"
+      pure $ response404 h
+  where
+    foldSets :: (Monad m) => Query -> Handle m -> [Handle m -> Query -> Handle m] -> Handle m
+    foldSets query = foldr (\set h' -> set h' query)
+
+    setPanigate :: (Monad m) => Handle m -> Query -> Handle m
+    setPanigate h' q =
+      let baseHandle = base h'
+          (userOffset, userLimit) = queryToPanigate q
+          newBaseHandle = baseHandle {Handlers.Base.Base.userOffset = userOffset, Handlers.Base.Base.userLimit = userLimit}
+       in h' {base = newBaseHandle}
+
+    setSort :: (Monad m) => Handle m -> Query -> Handle m
+    setSort h' q =
+      let baseHandle = base h'
+          (userSortColumn, userSortOrder) = queryToSort q
+          newBaseHandle = baseHandle {Handlers.Base.Base.sortColumnNews = userSortColumn, Handlers.Base.Base.sortOrderNews = userSortOrder}
+       in h' {base = newBaseHandle}
+
+    setFind :: (Monad m) => Handle m -> Query -> Handle m
+    setFind h' q =
+      let baseHandle = base h'
+          mbFind = queryToFind q
+          newBaseHandle = baseHandle {Handlers.Base.Base.findSubString = mbFind}
+       in h' {base = newBaseHandle}
+
+    setFilters :: (Monad m) => Handle m -> Query -> Handle m
+    setFilters h' q =
+      let baseHandle = base h'
+          filters = queryToFilters q
+          filterVisible = FilterPublishOrAuthor (fmap getLogin $ author $ client h) -- publish or author visible news
+          newBaseHandle = baseHandle {Handlers.Base.Base.filtersNews = filterVisible : filters}
+       in h' {base = newBaseHandle}
