+++ src/Web/WebType.hs
@@ -14,8 +14,9 @@ import qualified Data.Text as T
 import qualified Data.Text.Encoding as E
 import Data.Time (UTCTime)
 import GHC.Generics (Generic)
+import Handlers.Web.Web (NewsOut (..))
 import Scheme (Category (..), ColumnType (..), FilterItem (..), Find (..), Image (..), SortOrder (..), User (..))
-import Types (Content (..), Label (..), Name (..), NewsOut (..), Title (..), URI_Image (..))
+import Types (Content (..), Label (..), Name (..), Title (..), URI_Image (..))
 
 data UserToWeb = UserToWeb
   { name :: T.Text,
